recommender-ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-decouple==3.8
httpx==0.27.0
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.3
tqdm==4.66.1
tensorboard==2.15.1
matplotlib==3.8.2
seaborn==0.13.0
# CPU build by default. For NVIDIA GPU after pip install -r requirements.txt:
#   pip uninstall -y torch
#   pip install torch --index-url https://download.pytorch.org/whl/cu124
torch==2.2.2
